Reestrutura app e adiciona dashboard premium
