+++ verumoverview/frontend/src/App.tsx
@@ -1,0 +1,30 @@
+import { Routes, Route } from 'react-router-dom';
+import Layout from './components/layout/Layout';
+import Dashboard from './pages/Dashboard';
+import Projetos from './pages/Projetos';
+import Atividades from './pages/Atividades';
+import Pessoas from './pages/Pessoas';
+import Times from './pages/Times';
+import ControleAcesso from './pages/ControleAcesso';
+import Login from './pages/Login';
+import SolicitarAcesso from './pages/SolicitarAcesso';
+import DashboardPremium from './pages/DashboardPremium';
+
+export default function App() {
+  return (
+    <Routes>
+      <Route path="/login" element={<Login />} />
+      <Route path="/solicitar-acesso" element={<SolicitarAcesso />} />
+      <Route element={<Layout />}>
+        <Route index element={<Dashboard />} />
+        <Route path="dashboard" element={<Dashboard />} />
+        <Route path="premium" element={<DashboardPremium />} />
+        <Route path="projetos" element={<Projetos />} />
+        <Route path="atividades" element={<Atividades />} />
+        <Route path="pessoas" element={<Pessoas />} />
+        <Route path="times" element={<Times />} />
+        <Route path="controle-acesso" element={<ControleAcesso />} />
+      </Route>
+    </Routes>
+  );
+}
