+++ verumoverview/frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App.jsx';
+import App from './App';
 import { AuthProvider } from './hooks/AuthContext';
 import { ThemeProvider } from './hooks/ThemeContext';
 import { ToastProvider } from './hooks/ToastContext';
